Change the data read method

+++ app/src/main/java/e/gongfurui/digitallearnerlogbook/Roles/Learner.java
@@ -5,12 +5,12 @@ import java.util.ArrayList;
 public class Learner extends Role{
     public int driver_id;
     public String date_of_birth;
-    public int time;
+    public double time;
 
     public ArrayList<Boolean> courseProgressList;
     public ArrayList<String> courseCommentList;
 
-    public Learner(Role role, int driver_id, String date_of_birth, int time,
+    public Learner(Role role, int driver_id, String date_of_birth, double time,
                    ArrayList<Boolean> courseProgressList, ArrayList<String> courseCommentList){
         super(role.name, role.email, role.psw);
         this.driver_id = driver_id;
